PC-1121: move duplicate referral checker to its own file
